fastapi/requirements.txt
python-dotenv
numpy
pillow
requests
fastapi
uvicorn
transformers
accelerate
diffusers
torch

